+++ gemalaya/qml/LinksGroupItem.qml
@@ -10,7 +10,10 @@ GridLayout {
   columns: Conf.ui.links.gridColumns
   rowSpacing: Conf.ui.links.gridRowSpacing
   clip: Conf.ui.links.gridLimitHeight
+  opacity: collapsed ?
+           Conf.ui.links.gridCollapsedOpacityPercentage / 100 : 1
 
+  property bool collapsed: Conf.ui.links.gridLimitHeight && mHeight !== -1
   property real limitedHeight: Conf.ui.links.gridMaxCollapsedHeight
   property real mHeight: limitedHeight
   Layout.maximumHeight: Conf.ui.links.gridLimitHeight ? mHeight : -1
@@ -22,6 +25,17 @@ GridLayout {
 
   /* Skip the whole grid and focus the next thing in the page */
   signal skipGrid()
+
+  function expand() {
+    mHeight = limitedHeight
+  }
+
+  onFocusChanged: {
+    /* If the links grid gets the focus, focus the first link in the grid */
+
+    if (focus && children.length > 0)
+      children[0].focus = true
+  }
 
   onLinkFocused: {
     for (var idx=0; idx < children.length; idx++) {